Add env vars for compliance checking (#9)
This commit adds env vars for the certification, the definitions and the
path to the checks folder. The existence of these are confirmed at
startup by dotenv-safe based on the example values provided by
.env.example.

The createCompliance function is then used to create a root object for
the graphql api to expose.